Disable submit button on button click to avoid multiple ajax calls
 - Also, add test for the functionality
 - Bump version to 1.2.7

Summary: If submit button is clicked multiple times, multiple
ajax requests are sent. To avoid this, we disable the submit
button as soon as the submit button is clicked and enable it
again at the end of the ajax call depending on the response
from backend. This commit addresses EDUCATOR-273 issued in
openedx JIRA.